chore: simplify the test setup file
move some kubectl-specific content to kubectl.ts
